Add PrintService and move logic from facade to the service.
